implement rules
